* feature/redesign: (26 commits)
  Use the component name in the sidebar title (#11088)
  Redesign: layout center (#11068)
  Redesign: display only public users followings in following profile tab (#11051)
  Redesign: fix cards (#11072)
  Redesign: remove gcalendar warning (#10999)
  lock sass-embedded (#11077)
  Redesign: scopes picker (#11039)
  Redesign: pending notifications (#10926)
  Redesign: change the filters style (#11049)
  Redesign: secondary actions (#11067)
  Redesign: assembly members page (#10992)
  Redesign: update default static map size to 300px (#10953)
  Redesign: visual placeholder cards (#11038)
  Redesign: fix emoji popup position (#10957)
  Redesign: enable REDESIGN_ENABLED on ParticipatoryProcesses (#10983)
  Redesign: opinion buttons (#10998)
  Display a disabled message button in public profile if user blocks them instead of hidding it (#10984)
  Redesign: confirm modal Foundation-free (#10978)
  Redesign: initiatives (#10646)
  Redesign: surveys (#10922)
  ...

+++ decidim-core/app/cells/decidim/profile_actions_cell.rb
@@ -10,13 +10,15 @@ module Decidim
     ACTIONS_ITEMS = {
       edit_profile: { icon: "pencil-line", path: :profile_edit_path },
       create_user_group: { icon: "team-line", path: :profile_new_group_path },
+      resend_email_confirmation_instructions: { icon: "share-forward-line", path: :group_email_confirmation_path, options: { method: :post } },
       edit_user_group: { icon: "team-line", path: :edit_group_path },
       message: { icon: "mail-send-line", path: :new_conversation_path },
+      disabled_message: { icon: "mail-send-line", options: { html_options: { disabled: true, title: I18n.t("decidim.user_contact_disabled") } } },
       manage_user_group_users: { icon: "user-settings-line", path: :profile_group_members_path },
       manage_user_group_admins: { icon: "user-star-line", path: :profile_group_admins_path },
       invite_user: { icon: "user-add-line", path: :group_invites_path },
       join_user_group: { icon: "user-add-line", path: :group_join_requests_path, options: { method: :post } },
-      leave_user_group: { icon: "logout-box-r-line", path: :leave_group_path, options: { method: :delete } }
+      leave_user_group: { icon: "logout-box-r-line", path: :leave_group_path, options: { method: :delete, data: { confirm: I18n.t("decidim.groups.actions.are_you_sure") } } }
     }.freeze
 
     private
@@ -61,7 +63,7 @@ module Decidim
       @actions_keys ||= [].tap do |keys|
         keys << :edit_profile if own_profile?
         keys << :create_user_group if own_profile? && user_groups_enabled?
-        keys << :message if can_contact_user?
+        keys << message_key if can_contact_user?
         keys << :join_user_group if can_join_user_group?
         keys << :leave_user_group if can_leave_group?
       end
@@ -75,12 +77,19 @@ module Decidim
                                          :manage_user_group_admins,
                                          :invite_user
                                        ].tap do |keys|
+                                         keys.prepend(:resend_email_confirmation_instructions) if user_group_email_to_be_confirmed?
                                          keys << :join_user_group if can_join_user_group?
                                          keys << :leave_user_group if can_leave_group?
                                        end
                                      else
                                        []
                                      end
+    end
+
+    def message_key
+      return :message if current_or_new_conversation_path_with(presented_profile).present?
+
+      :disabled_message
     end
 
     def profile_actions
@@ -117,12 +126,19 @@ module Decidim
       !Decidim::UserGroupMembership.exists?(user: current_user, user_group: model)
     end
 
+    def user_group_email_to_be_confirmed?
+      return false unless user_group?
+      return false unless current_user
+
+      !model.confirmed?
+    end
+
     def group_member?
       Decidim::UserGroupMembership.exists?(user: current_user, user_group: model)
     end
 
     def can_contact_user?
-      !own_profile? && presented_profile.can_be_contacted? && current_or_new_conversation_path_with(presented_profile).present?
+      !current_user || (current_user && current_user != model && presented_profile.can_be_contacted?)
     end
   end
 end
